requirements.txt
aioredis==1.3.1
amqp==5.0.9
arabic-reshaper==2.1.3
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
autobahn==21.11.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.2.1
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.8
django-allauth==0.45.0
django-autoslug==1.9.8
django-braces==1.15.0
django-celery-results==2.2.0
django-ckeditor==6.1.0
django-cleanup==5.2.0
django-crispy-forms==1.13.0
django-environ==0.8.1
django-filter==21.1
django-formtools==2.3
django-js-asset==1.2.2
django-model-utils==4.2.0
django-phonenumber-field==5.2.0
django-redis==5.2.0
django-star-ratings==0.9.2
django-utils-six==2.0
future==0.18.2
hiredis==2.0.0
html5lib==1.1
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
jsonfield==3.1.0
kombu==5.2.2
msgpack==1.0.3
oauthlib==3.1.1
phonenumbers==8.12.35
Pillow==8.3.2
prompt-toolkit==3.0.24
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==2.2.0
pyOpenSSL==21.0.0
PyPDF2==1.26.0
python-bidi==0.4.2
python3-openid==3.2.0
pytz==2021.3
redis==3.5.3
reportlab==3.6.6
requests==2.26.0
requests-oauthlib==1.3.0
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.2
swapper==1.3.0
twilio==7.7.0
Twisted==21.7.0
twisted-iocpsupport==1.0.2
txaio==21.2.1
typing_extensions==4.0.1
urllib3==1.26.7
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
xhtml2pdf==0.2.5
zope.interface==5.4.0
